graphics class
code from playcontroller go to graphics
